10/requirements.txt
json
ujson
setuptools
time